django/coolsite/requirements.txt
Django>=5.0,<6.0
gunicorn
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
djoser
django-captcha
django-debug-toolbar
django-simple-captcha

